+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 ## v4.2.2
 
 ### Improvements
-- /packcore diagnose now include screen resolution.
+- `/packcore diagnose` now include screen resolution.
+- Added `/packore config` to open the config menu.